Create Question142.cpp
142 Given the head of a doubly linked lis. The values of the linked list are sorted 
in non-decreasing order. Find if there exists a pair of distinct nodes such that the
sum of their values is x. Return the pair in the from of a vector[I r],Where I & r are
the values stored in the pair 2nodes pointed by the pointers. If there are multiple such pain,
return any of them. If there is no such pair return[-1, -1]